[FAB-5430] Initial commit in the fabric-test repo
This is the initial commit for the fabric-test
repository containing automated tests and tools.

This is a port from the fabric repository with the
exception of changing a couple of files in feature/steps
to use the new directory structure to ensure that tests
will run.

There are also the removal of a few superfluous files.

Change-Id: I0a3997eda8060180977119c3817e12674fb0e3cb
Signed-off-by: Latitia M Haskins <latitia.haskins@gmail.com>