[FAB-5320]Decrease total run time of Auction Tests
Removed the RUN_TIME_HOURS parameter which was added to run the auction tests for 1 hr.
As a result, all the auction tests are executed only one time.

Removed the tests which were running as part of the 1 hr loop and added them separately as new tests.

Decreased the total runtime to < 10 min (Test run was for 355 seconds)

https://gerrit.hyperledger.org/r/#/c/11711/

[FAB-5608] Fix Daily Auction chaincode tests

Auction chaincode tests use the configtx.yaml from e2e cli example.
Recently, configtx.yaml was updated in e2e cli to use 'kafka' orderer.

This patch adds a fix to configtx.yaml to use 'solo' orderer for running auction tests.

https://gerrit.hyperledger.org/r/#/c/12275/

Change-Id: I78f2c68f96f5df9f04074e33e651a19ba2041567
Signed-off-by: sandp125 <sandeep.pulluru@itpeoplecorp.com>