[FAB-5480] Refactor allowing different Interfaces
This changes allows for tests to be executed using
different interfaces when interacting with the fabric
network. It starts with cli, sdk, and tools.

This is the first of numerous CRs for this change. For
now, the default will be to use the CLI interface, until
the changes for SDK are fully implemented.

Change-Id: I1c8d622594c278a25c36e883647724a354536857
Signed-off-by: Latitia M Haskins <latitia.haskins@gmail.com>