[FAB-536,FAB-3634] Add SDK and Remote Support
This adds code that will allow executing behave tests
using different types of interface - whether a tool
that executes chaincode or different SDKs.

The first SDK that is supported with this change is
the node SDK.

This also adds code that will allow executing behave tests
on an existing network. There is some setup that is
necessary in order for the test to know what nodes
should to mapped to what location.

See the Remote Execution section in the README for
more details on how this can be done.

Change-Id: I58bfd158706bce8b4e1c9b7b143bb7f8f0d21aa1
Signed-off-by: Latitia M Haskins <latitia.haskins@gmail.com>