WIP:[FAB-3634] Allow executing on a remote network
This adds code that will allow eecuting behave tests
on an existing network. There is some setup that is
necessary in order for the test to know what nodes
should to mapped to what location.

See the Remote Execution section in the README for
more details on how this can be done.

Change-Id: I58bfd158706bce8b4e1c9b7b143bb7f8f0d21aa1
Signed-off-by: Latitia M Haskins <latitia.haskins@gmail.com>