[FAB-6665] fabric-test runSmokeTestSuite.sh
This script runSmokeTestSuite.sh can be called
by Jenkins or the fabric-test/Makefile
to run the existing Behave smoke tests.
This update also implements the high level
mechanism how to invoke the pte tests.
For now, the performance tests in systest_pte.py
are still WIP and are tagged to be skipped; they
can be enabled after being finalized when we
merge the testcase scripts that they utilize.

Change-Id: Ia339e1a934dbf149d71e73c0d7c2d971c50b3bb1
Signed-off-by: Scott Zwierzynski <scottz@us.ibm.com>