[FAB-6665] fabric-test runSmokeTestSuite.sh
The new script runSmokeTestSuite.sh can be called
to run the existing Behave smoke tests
on command line, or by Jenkins, or from the
fabric-test/Makefile that is being published soon.
This update also implements the high level
mechanism how to invoke the pte systests.
For now, the performance tests in systest_pte.py
are still WIP and thus tagged to be skipped; they
can be enabled after being finalized when we
merge the testcase scripts that they utilize.

Change-Id: Ia339e1a934dbf149d71e73c0d7c2d971c50b3bb1
Signed-off-by: Scott Zwierzynski <scottz@us.ibm.com>