[FAB-2212] Orderer Traffic Engine initial commit
Orderer Traffic Engine (OTE) tool tests the operation of a
hyperledger fabric ordering service. OTE sends transactions to
every channel on every orderer, and verifies that the correct number
of transactions and blocks are delivered on every channel from every orderer.
OTE stresses the ordering service by sending large number of concurrent
transactions when specified.

Producer clients are created to connect via grpc ports to the orderers to
concurrently send traffic until the requested number of transactions are sent.
Each client generates unique transactions - a fraction of the total
requested number of transactions.

Consumer clients are created to connect via grpc ports to the orderers
to concurrently receive delivered traffic until all batches of transactions
are tallied. OTE checks if the correct number of blocks and TXs are delivered
by all the orderers on all the channels.

OTE generates report logs and returns a pass/fail boolean and a resultSummaryString

Change-Id: Ic2a07bfbf2da43d6d7036d33fdcd85323563956e
Signed-off-by: Surya <suryalnvs@gmail.com>