[FAB-6500] Refactor code for fabric-ca inclusion
This is some refactor work that is needed as a base for
including the fabric-ca as a possible bootstrap process.

This includes moving the enableTLS call to the
compose utility instead of in the common utility.
Adds a method for retrieving environment variables
from a stated container.
Adds user parameter to the calls.
Adds orderer host names to the docker composition files.

Change-Id: I088edfa6eb2b2d77a2933799d570d5711d947110
Signed-off-by: Latitia M Haskins <latitia.haskins@gmail.com>