[FAB-5548] update LTE running location
This CR updates the LTE so that it copies itself into the fabric
directory before executing the tests. This is to circumvent vendoring
issues. Also, for now, it removes the existing LTE in fabric repo before
copying sand running tests.

Also adds some fix that was only available in the fabric version of LTE.

Change-Id: I2f9ab216b8502d1580eba67dbe46208241b69901
Signed-off-by: Adnan Choudhury <adnan.choudhury@itpeoplecorp.com>