[FAB-6387] Fix chaincode negative tests
Test case 4720/4721/4722 assumed that error code is 400. But in
these "chaincode not found error cases", error code 500 is more
appropriate. Also 4721/4722 expect a success status.
This CR changes error code from 400 to 500 and response type from
success to error. In future release v1.2, there are plans to set
the error code to 404, but for now, agreement was reached for the
current code to return 500.

Change-Id: I9bf0787952fbd5eadfc70a30a894e925e9501246
Signed-off-by: Nao Nishijima <nao.nishijima@hal.hitachi.com>