FAB-7408/FAB-7409:Add impl statements and fetch calls
This adds some new implementation statements for
updating channel configuration updates for a fabric
network.

Adds fetch calls for obtaining the channel config
for updating a channel.

Updates the submodules to the latest version to
ensure that any supporting code (e.g. chaincodes)
or configs that are being used in this suite of
tests is current.

Change-Id: I4dd50724acfcd2a0acd7e4907c9dbe1fed944405
Signed-off-by: Latitia M Haskins <latitia.haskins@gmail.com>