[FAB-10535] Changes to ATD to check pte k8s
ATD to check the status of PTE pod. ATD will wait until the pte status
is changed to completed and then places the logs under ATD/vars directory
with testcase name in its name.

Using configMaps for testcase script and mounting configMap to the pod
and use testcases script as an entrypoint in the pte pod instead of sleep

Change-Id: I26330857def0fe1d305b61c14099a50c088b592e
Signed-off-by: Surya <suryalnvs@gmail.com>