[FAB-10781] Release-1.1 CI build process
The git submodules should not be updated with what
is in master, but should be tested at the specified
level. If the commit level needs to change, then
the submodule commit should change.

This also makes the necessary changes in order to
get the tests to execute successfully.

Also has PTE access the release-1.1 release.

Change-Id: I4cfd97685637463d219f948502af7d7acff67b3d
Signed-off-by: Latitia M Haskins <latitia.haskins@gmail.com>