[FAB-11491] Support for custom org/MSP names in NL
Added option to pass a JSON file as a parameter to the scripts in this
folder.
The JSON file contains default organization names and MSP IDs as keys,
and custom names and IDs as values.
Channel artifacts, cryptographic material, and docker-compose files are
created using custom names if a file is provided as a parameter.
If no mapping is specified for a given organization name or MSP ID in
the file or if no file is specified, the default names and IDs are used.

Change-Id: Idb5385218ca927950bbcb7e46e8896fd268c111f
Signed-off-by: vramakr2 <vramakr2@in.ibm.com>