[FAB-11492] Chaincode Access Policy Spec for PTE
Added a feature to allow a user to specify which organizations'
identities will be used to call which chaincode function (query or
invoke) to meet the access rules embedded within chaincode.
An admin user is enrolled for every organization in the network a
priori before starting the traffic.
When a given transaction is to be executed, the list of permitted
organizations is looked up, and a user handle corresponding to one of
them is associated with the fabric client instance that is used to
trigger the transaction.

Change-Id: I26431c48843d8b66b52a08b96e3f074c4b9b0d68
Signed-off-by: vramakr2 <vramakr2@in.ibm.com>