[FAB-11637] fix subprocess error handling
The PTE test suite uses python subprocess calls
to execute test steps. An exit code of 1 (failure)
can be returned when the command or args are not
formatted correctly. It also returns 1 for any
grep command that returns zero result lines, even
when the command does complete successfully; that
means a CallProcessError is raised instead of
simply an assertion error when the zero count
does not match the asserted value. This fix
pipes the grep result to the "wc -l", which will
return a successful (zero) exit code and then
execute the assertion statement as intended.

This fix also allows for the case where the
expected count from the grep is zero, whereas
the prior behavior would result in an error.

Change-Id: I56c135b9c7466460d8012d331a00d98fd28218c7
Signed-off-by: Scott Zwierzynski <scottz@us.ibm.com>