[FAB-11898] archive data for Testviewer
Consistently archive the missing daily test
results for PTE, OTE, and LTE tests.
Also fix the TestViewer tool to
retrieve test data (up to 6 months) from Nexus,
now that it is archived there after recent
changes in the jobs, instead of from jenkins,
where it is saved only 1 week. We do this for
the PTE and OTE and LTE tests.

Change-Id: I9cb34ce090e932f697c82d7c8ee247f65da58021
Signed-off-by: Scott Zwierzynski <scottz@us.ibm.com>