FAB-11978 Pull the javaenv for smoke & daily tests
This change is to build the javaenv image in the
ci-smoke and ci-daily test targets; ci-smoke will
be used by the verify and merge jobs in fabric-test
repo. And we enhance pullDockerImages.sh so the pull-images
test target will also pull the javaenv image from Nexus;
pull-images will be used by the daily svt jobs.

Change-Id: Iab6f88b2a8f18f0265ab27e1248c22963de9b0ea
Signed-off-by: Sambhav Nidamarty <sambhavdutt@gmail.com>