[FAB-11999] Fix a query to get a correct response
FAB-7264 and FAB-7269 behave tests failed to get a "no_usable_index"
response. A part of these tests expects query failures. However, the
query return success. The reason is that the query is looking for
the key with an index.
This CR fixes an argument to get a query failure.

Change-Id: Id40074331e97709a1505848bb1cf0dee2f565932
Signed-off-by: Nao Nishijima <nao.nishijima@hal.hitachi.com>