[FAB-9624] PTE Support for Poisson Traffic
Replaced constant rate traffic generators with generic functions that can
support constant (uniform) or any other distribution.
User must implement a function to compute a backoff before the next
transaction in every iteration, according to the desired probability distribution
(or schedule.)
A handle to this function is passed to the generic traffic generation
function.
Added a Poisson traffic generator (backoff following an exponential
distribution) for illustration.
Added support for an optional function to support deviations from the
base distribution/schedule; this is to support the `devFreq` parameter
in `constantOpts`.

Updated the README with instructions.

Change-Id: I413fc0ae7bfa6a15618a63a7e7eb1f134c4c2277
Signed-off-by: vramakr2 <vramakr2@in.ibm.com>