[FAB-16265] Retire unneeded Behave tests, part 1
Note: this patchset2 contains a change to
force the behave daily tests to be run
as part of the verify job, so we expect it
to take possibly 3-6 hours. The code will
be removed/reverted to patchset1 (and
remove this comment in this commit message)
before we merge.

The tests that we are disabling or removing
are either duplicated already in
fabric UT/IT or are deemed unnecessary.
(The remaining tests will be disabled and
removed after they are replaced.)
FAB-16265 #done
FAB-16266 #done

Change-Id: I236c9318b7b417b10fbe6bcbf7cee5027f9e425e
Signed-off-by: Scott Zwierzynski <scottz@us.ibm.com>